:recycle: refactor(#168/mandalart): mandalart 컨트롤러 관련 클래스 ninedot-api 모듈로
