+++ ninedot-api/src/main/java/org/sopt36/ninedotserver/mandalart/v1/CoreGoalController.java
@@ -1,20 +1,20 @@
-package org.sopt36.ninedotserver.mandalart.controller;
+package org.sopt36.ninedotserver.mandalart.v1;
 
-import static org.sopt36.ninedotserver.mandalart.controller.message.CoreGoalMessage.AI_RESPONSE_SUCCESS;
-import static org.sopt36.ninedotserver.mandalart.controller.message.CoreGoalMessage.CORE_GOAL_AI_CREATED_SUCCESS;
-import static org.sopt36.ninedotserver.mandalart.controller.message.CoreGoalMessage.CORE_GOAL_CREATED_SUCCESS;
-import static org.sopt36.ninedotserver.mandalart.controller.message.CoreGoalMessage.CORE_GOAL_IDS_RETRIEVED_SUCCESS;
-import static org.sopt36.ninedotserver.mandalart.controller.message.CoreGoalMessage.CORE_GOAL_LIST_RETRIEVED_SUCCESS;
-import static org.sopt36.ninedotserver.mandalart.controller.message.CoreGoalMessage.CORE_GOAL_ONBOARDING_DELETED_SUCCESS;
-import static org.sopt36.ninedotserver.mandalart.controller.message.CoreGoalMessage.CORE_GOAL_ONBOARDING_UPDATED_SUCCESS;
-import static org.sopt36.ninedotserver.mandalart.controller.message.CoreGoalMessage.CORE_GOAL_UPDATED_WITH_SUB_GOALS_SUCCESS;
+import static org.sopt36.ninedotserver.mandalart.v1.message.CoreGoalMessage.AI_RESPONSE_SUCCESS;
+import static org.sopt36.ninedotserver.mandalart.v1.message.CoreGoalMessage.CORE_GOAL_AI_CREATED_SUCCESS;
+import static org.sopt36.ninedotserver.mandalart.v1.message.CoreGoalMessage.CORE_GOAL_CREATED_SUCCESS;
+import static org.sopt36.ninedotserver.mandalart.v1.message.CoreGoalMessage.CORE_GOAL_IDS_RETRIEVED_SUCCESS;
+import static org.sopt36.ninedotserver.mandalart.v1.message.CoreGoalMessage.CORE_GOAL_LIST_RETRIEVED_SUCCESS;
+import static org.sopt36.ninedotserver.mandalart.v1.message.CoreGoalMessage.CORE_GOAL_ONBOARDING_DELETED_SUCCESS;
+import static org.sopt36.ninedotserver.mandalart.v1.message.CoreGoalMessage.CORE_GOAL_ONBOARDING_UPDATED_SUCCESS;
+import static org.sopt36.ninedotserver.mandalart.v1.message.CoreGoalMessage.CORE_GOAL_UPDATED_WITH_SUB_GOALS_SUCCESS;
 
 import jakarta.validation.Valid;
 import java.net.URI;
 import lombok.RequiredArgsConstructor;
 import org.sopt36.ninedotserver.ai.dto.response.CoreGoalAiResponse;
-import org.sopt36.ninedotserver.ai.service.AiRecommendationService;
-import org.sopt36.ninedotserver.global.dto.response.ApiResponse;
+import org.sopt36.ninedotserver.ai.usecase.AiRecommendationService;
+import org.sopt36.ninedotserver.dto.response.ApiResponse;
 import org.sopt36.ninedotserver.mandalart.dto.request.CoreGoalAiCreateRequest;
 import org.sopt36.ninedotserver.mandalart.dto.request.CoreGoalCreateRequest;
 import org.sopt36.ninedotserver.mandalart.dto.request.CoreGoalUpdateRequest;
@@ -24,8 +24,8 @@ import org.sopt36.ninedotserver.mandalart.dto.response.CoreGoalAiListResponse;
 import org.sopt36.ninedotserver.mandalart.dto.response.CoreGoalCreateResponse;
 import org.sopt36.ninedotserver.mandalart.dto.response.CoreGoalIdsResponse;
 import org.sopt36.ninedotserver.mandalart.dto.response.CoreGoalsResponse;
-import org.sopt36.ninedotserver.mandalart.service.command.CoreGoalCommandService;
-import org.sopt36.ninedotserver.mandalart.service.query.CoreGoalQueryService;
+import org.sopt36.ninedotserver.mandalart.usecase.command.CoreGoalCommandService;
+import org.sopt36.ninedotserver.mandalart.usecase.query.CoreGoalQueryService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.DeleteMapping;
