+++ ninedot-api/src/main/java/org/sopt36/ninedotserver/mandalart/v1/RecommendationController.java
@@ -1,14 +1,14 @@
-package org.sopt36.ninedotserver.mandalart.controller;
+package org.sopt36.ninedotserver.mandalart.v1;
 
-import static org.sopt36.ninedotserver.mandalart.controller.message.RecommendationMessage.RECOMMENDATION_CREATED_SUCCESS;
-import static org.sopt36.ninedotserver.mandalart.controller.message.RecommendationMessage.RECOMMENDATION_RETRIEVED_SUCCESS;
+import static org.sopt36.ninedotserver.mandalart.v1.message.RecommendationMessage.RECOMMENDATION_CREATED_SUCCESS;
+import static org.sopt36.ninedotserver.mandalart.v1.message.RecommendationMessage.RECOMMENDATION_RETRIEVED_SUCCESS;
 
 import java.time.LocalDate;
 import lombok.RequiredArgsConstructor;
-import org.sopt36.ninedotserver.global.dto.response.ApiResponse;
+import org.sopt36.ninedotserver.dto.response.ApiResponse;
 import org.sopt36.ninedotserver.mandalart.dto.response.SubGoalListResponse;
-import org.sopt36.ninedotserver.mandalart.service.command.RecommendationSchedulerService;
-import org.sopt36.ninedotserver.mandalart.service.query.RecommendationQueryService;
+import org.sopt36.ninedotserver.mandalart.usecase.command.RecommendationSchedulerService;
+import org.sopt36.ninedotserver.mandalart.usecase.query.RecommendationQueryService;
 import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
