+++ ninedot-api/src/main/java/org/sopt36/ninedotserver/mandalart/v1/SubGoalController.java
@@ -1,12 +1,12 @@
-package org.sopt36.ninedotserver.mandalart.controller;
+package org.sopt36.ninedotserver.mandalart.v1;
 
-import static org.sopt36.ninedotserver.mandalart.controller.message.SubGoalMessage.SUB_GOAL_AI_CREATED_SUCCESS;
-import static org.sopt36.ninedotserver.mandalart.controller.message.SubGoalMessage.SUB_GOAL_AI_RECOMMENDATION_SUCCESS;
-import static org.sopt36.ninedotserver.mandalart.controller.message.SubGoalMessage.SUB_GOAL_CREATE_SUCCESS;
-import static org.sopt36.ninedotserver.mandalart.controller.message.SubGoalMessage.SUB_GOAL_DELETE_SUCCESS;
-import static org.sopt36.ninedotserver.mandalart.controller.message.SubGoalMessage.SUB_GOAL_ID_LIST_FETCH_SUCCESS;
-import static org.sopt36.ninedotserver.mandalart.controller.message.SubGoalMessage.SUB_GOAL_READ_SUCCESS;
-import static org.sopt36.ninedotserver.mandalart.controller.message.SubGoalMessage.SUB_GOAL_UPDATE_SUCCESS;
+import static org.sopt36.ninedotserver.mandalart.v1.message.SubGoalMessage.SUB_GOAL_AI_CREATED_SUCCESS;
+import static org.sopt36.ninedotserver.mandalart.v1.message.SubGoalMessage.SUB_GOAL_AI_RECOMMENDATION_SUCCESS;
+import static org.sopt36.ninedotserver.mandalart.v1.message.SubGoalMessage.SUB_GOAL_CREATE_SUCCESS;
+import static org.sopt36.ninedotserver.mandalart.v1.message.SubGoalMessage.SUB_GOAL_DELETE_SUCCESS;
+import static org.sopt36.ninedotserver.mandalart.v1.message.SubGoalMessage.SUB_GOAL_ID_LIST_FETCH_SUCCESS;
+import static org.sopt36.ninedotserver.mandalart.v1.message.SubGoalMessage.SUB_GOAL_READ_SUCCESS;
+import static org.sopt36.ninedotserver.mandalart.v1.message.SubGoalMessage.SUB_GOAL_UPDATE_SUCCESS;
 
 import jakarta.validation.Valid;
 import java.net.URI;
@@ -15,9 +15,9 @@ import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.sopt36.ninedotserver.ai.dto.request.SubGoalAiRequest;
 import org.sopt36.ninedotserver.ai.dto.response.SubGoalAiResponse;
-import org.sopt36.ninedotserver.ai.service.AiSubGoalRecommendationService;
-import org.sopt36.ninedotserver.global.dto.response.ApiResponse;
-import org.sopt36.ninedotserver.mandalart.domain.Cycle;
+import org.sopt36.ninedotserver.ai.usecase.AiSubGoalRecommendationService;
+import org.sopt36.ninedotserver.dto.response.ApiResponse;
+import org.sopt36.ninedotserver.mandalart.model.Cycle;
 import org.sopt36.ninedotserver.mandalart.dto.request.SubGoalAiListCreateRequest;
 import org.sopt36.ninedotserver.mandalart.dto.request.SubGoalCreateRequest;
 import org.sopt36.ninedotserver.mandalart.dto.request.SubGoalUpdateRequest;
@@ -26,8 +26,8 @@ import org.sopt36.ninedotserver.mandalart.dto.response.SubGoalCreateResponse;
 import org.sopt36.ninedotserver.mandalart.dto.response.SubGoalIdListResponse;
 import org.sopt36.ninedotserver.mandalart.dto.response.SubGoalIdResponse;
 import org.sopt36.ninedotserver.mandalart.dto.response.SubGoalListResponse;
-import org.sopt36.ninedotserver.mandalart.service.command.SubGoalCommandService;
-import org.sopt36.ninedotserver.mandalart.service.query.SubGoalQueryService;
+import org.sopt36.ninedotserver.mandalart.usecase.command.SubGoalCommandService;
+import org.sopt36.ninedotserver.mandalart.usecase.query.SubGoalQueryService;
 import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
